+++ docs/changes.md
@@ -1,4 +1,8 @@
 Changes of `joker-xopen`
 ------------------------
 
+### 0.2
+* use joker-minions >= 0.1 (with verbs like del, get, set, etc.)
+* xopen verbs without leading '#'
+* -g, --get option